requirements.txt
numpy
Pillow
tensorflow
tensorflow-hub
google-cloud-storage
google-cloud-firestore